[Merge Redux] Extract helper/utility functions
